Extract stub arg-binding materialization into its own module
serialized_objects.py was carrying the entire TaskFlow arg-binding builder
inline, pulling in pydantic schema generation and several SDK-only imports
for logic that only fires for @task.stub tasks. Splitting it into
stub_arg_bindings.py keeps the core serialization module's import surface
focused, and a dedicated arg_bindings property replaces direct access to
the private _arg_bindings attribute from outside SerializedBaseOperator.

The builder is imported lazily too, so Python-only deployments never pay for
pydantic's JSON-schema machinery just to serialize a Dag. Review also caught
that schema.json described the binding in the {__type, __var} encoded form the
serializer never emits, and that the test asserting it validated nothing --
schema.json's tasks array never reaches the operator subschema, so the
definition is now asserted directly instead.
